requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
django==3.2.6
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
pillow==8.3.1
psycopg2-binary==2.9.1
pyjwt==2.1.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
